Add SQL query editor, home screen, specta type codegen
- SQL editor with CodeMirror 6 (PostgreSQL syntax highlighting,
  schema-aware autocomplete), query execution, results grid, and
  execution stats
- Home screen: clickable logo navigates to a connections overview
  with cards for each saved connection
- Migrate frontend/backend types to specta + tauri-specta codegen
  (bindings.ts), eliminating hand-mirrored TS types. Row/cell data
  crosses the IPC boundary as JSON-encoded strings to work around a
  known specta recursion limitation with serde_json::Value
- Normalize shadcn component import paths after a manual file
  reorganization (lib/, hooks/ moved under src/)
- 9 new integration tests for query execution against real Postgres

+++ src-tauri/src/domain/schema/models.rs
@@ -1,15 +1,16 @@
 use serde::Serialize;
+use specta::Type;
 
-#[derive(Debug, Clone, Serialize)]
+#[derive(Debug, Clone, Serialize, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct TableRef {
     pub schema: String,
     pub name: String,
     pub kind: String,
-    pub estimated_rows: i64,
+    pub estimated_rows: f64,
 }
 
-#[derive(Debug, Clone, Serialize)]
+#[derive(Debug, Clone, Serialize, Type)]
 #[serde(rename_all = "camelCase")]
 pub struct ColumnInfo {
     pub name: String,
